small-fix

+++ src/services/SlideService.ts
@@ -1,4 +1,9 @@
 import SlideRepository from '../repositories/SlideRepository'
-const deletePresentation = async (id: string) =>
+import GenericError from '../helpers/GenericError'
+
+const deleteSlide = async (id: string) => {
+  const slideById = await SlideRepository.findSlide(id)
+  if (!slideById) throw new GenericError('slide is not exist')
   await SlideRepository.deleteSlide(id)
-export default { deletePresentation }
+}
+export default { deleteSlide }
